+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - :star2: Docs & Docs & Docs, lots of new & better documentation
 - :star2: Added Support to the ChainBuilder for the `Split` operation
 - :star2: Added ne Filter operation to filter data.
-
+- :star2: Feature #17 - Add new complex item MixItem. This allows operations to return more than one type of item.
 
 # 1.0.3
 - :wrench: Fix - Stop Item is sent multiple times when using a step returns multiple GroupedItem's